Make RateLimitedRepeater private
See https://github.com/SnabbCo/snabbswitch/pull/793.
